Integration push July 19

+++ Customer/Gemini-Chef.sh
@@ -6,16 +6,19 @@ echo    # (optional) move to a new line
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
 #  echo "Login to the Gemini Repo"
-#  if !(docker login https://secure-registry.gsintlab.com)
+#  if !(docker login https://registry.gemini-systems.net/)
 #  then
 #	echo "Try Running Again with Corrrect Login Credentials. ERROR: Login Failed...Exiting..."
 #	exit
 #  fi         
-  echo "Pull Chef Server from Internal Registry..."
-  docker pull secure-registry.gsintlab.com/gemini/gemini-chef
+  echo "Pull Chef Server from Registry..."
+  docker pull registry.gemini-systems.net/gemini/gemini-chef
   echo "Continue to run chef ..."
-  docker rm -f gemini-chef
+  if docker ps -a |grep -aq gemini-chef; then
+        docker rm -f gemini-chef
+  fi
+
   ip=`curl -s http://ipecho.net/plain; echo`
-  docker run -it -p 443:443 --privileged -v /etc/chef-server/ --name gemini-chef -h $ip -d secure-registry.gsintlab.com/gemini/gemini-chef
+  docker run -it -p 443:443 --privileged -v /etc/chef-server/ --name gemini-chef -h $ip -d registry.gemini-systems.net/gemini/gemini-chef
 fi
